Fix duplicate playlist entries causing abort in rb_insert().
If a file appears multiple times on the command line or a playlist
containing a file appears before the file itself, then that file
gets added to the playlist multiple times.  An attempt is then made
to add the file to the rb-tree a second time when building the menu.