C99: Use AF_UNIX for both address and protocol families.
There is some confusion over whether AF_UNIX or AF_LOCAL is the now
accepted constant to use for the address family, and whether PF_*
or AF_* should be used for the protocol family.  As far as can be
determined from the POSIX standards documents available they seem to
have settled on AF_UNIX [1], but it seems counter-intuitive that an
OS-neutral standard would opt for that rather than the more generic
(and descriptive) AF_LOCAL.

The GNU C Library documentation says (in the section "Address Formats")
that AF_UNIX is a synonym for AF_LOCAL [2] and that the latter is
mandated by POSIX.1g while the unix(7) manpage suggests the opposite
[3].

It also seems generally agreed that all PF_* are now numerically equal
to their AF_* counterparts, using AF_* for protocol is correct and
that breaking this equality would also break a lot of existing code
[4].  The socket(3) manpage says [5]:

    "... subsequent standards use AF_* everywhere."

However, that would seem to imply that there can only be one protocol
per address family (unless non-address AF_* constants are introduced
to confuse the issue even further) in which case the protocol family
is completely redundant.

So whatever the standards, headers and commentators may say, AF_UNIX
is the one which seems to work in all cases tested so far.

(The OSX header file [6] is actually broken with respect to PF_*.
They are #define'd in a chain from AF_UNIX, but the chain is broken
by an #ifdef at AF_LOCAL for POSIX.)

Refs: [1] http://pubs.opengroup.org/onlinepubs/009695399/basedefs/sys/socket.h.html
      [2] http://www.gnu.org/software/libc/manual/html_node/Address-Formats.html
      [3] http://manpages.org/unix/7
      [4] http://www.kohala.com/start/lanciani.96apr10.txt
      [5] http://manpages.org/socket/2
      [6] http://opensource.apple.com/source/xnu/xnu-1456.1.26/bsd/sys/socket.h