FFmpeg: Circumvent 1MiB memory leak on opening empty or corrupt files.
The leak was isolated to a single FFmpeg commit (80a3a66) which
hinted strongly at the cause.  It pointed to code which, when read,
revealed that FFmpeg allocates memory in av_probe_input_buffer2() for
a probe buffer which is (subsequent to that commit) never released as
the result of attempting to reuse the probed data in later decoding
(which obviously won't happen if FFmpeg failed to identify the format).

This circumvention tracks the address of the AVIOContext block
containing the pointer to the buffer and frees both if FFmpeg hasn't.

Obviously, neither avio_close() or a single av_free() cannot be used
in this circumstance as the first segfaults and the second would
only free the AVIOContext block.  Ideally, FFmpeg would provide a
avio_free_context() function to allow this clean up to be done
atomically and completely ('cause only FFmpeg knows what memory it
might have allocated) if avformat_open_input() fails.

This patch has been tested against suitably corrupt files, but a
similarly corrupt audio stream is almost a contradiction in terms.

Also see: r2880
Ref: https://trac.ffmpeg.org/ticket/5584