Fix: Build failure on split ncurses/tinfo systems.
At the time of this commit, there is a bug in AX_WITH_CURSES which
causes it to test a stale value in 'pkg_failed' and fail even if the
ncurses .pc-file(s) were found.  To circumvent this bug, we unset the
'pkg_failed' shell variable before calling AX_WITH_CURSES.  This
unset can be removed once the macro has been fixed.

The ncurses .pc-file(s) define '_GNU_SOURCE' in Cflags.  This causes
problems for MOC when it is defined mid-configure, so we remove it.
It seems there's no reason to set it for client applications (using
the core ncurses header file as MOC does, at least), so ncurses should
not inflict it on all the clients component files.

The way in which 'configure' was unpacking CURSES_LIBS to test for
the presence of the set_escdelay() function no longer works now multiple
flags are present, so instead we fake it by asking for a check against
libm (assumed to always be present) but include all the CURSES_LIBS
flags as "other libraries" to be linked and therefore satisfying (or
not) the test.

Resolves: node/1469 (fix build on systems with split ncurses/tinfo)