Clean: Relocate SIGCHLD signal handler into the server.
Only the server process of MOC spawns child processes, so the more
logical place for the SIGCHLD handler is in the server code.  This
relocation also fixes any possibility of unreaped children when the
server is run in foreground mode.

Thanks to: Daniel T. Borelli (on whose patch this was based)
Also see: node/1554