Fix: Build failure on MacOS 10.12 with clock_gettime(2).
As of Darwin 16.0.0 Apple now supports clock_gettime(2) but does so in
a way which makes testing for it quite difficult.  This results in MOC
thinking it is present when it isn't or thinking it isn't when it is.
The problem affects many open source applications, not just MOC.

The most reliable way to resolve its presence is to just test on the
Darwin release number which, although not ideal, works consistantly
for MOC.

Resolves: node/1576 (2.6-alpha3 build failure on macOS 10.12)