Maint: Address FFmpeg/LibAV's deprecation of AVStream.codec.
There is no guidance in the documentation of a migration path
for AVStream.codec and it's obvious that the FFmpeg maintainers
do not have a solution either.  The most consistant narrative
which can be pieced together from the scattered clues is that
the management of the AVCodecContext struct should be brought
in-house and MOC should create and manage its own AVCodecContext
struct in all cases.

This may break when the API goes off on yet another tangent.