Maint: Use AVERROR_EOF to indicate EOF in FFmpeg/LibAV read callback.
FFmpeg/LibAV has recently documented the use of AVERROR_EOF to indicate
end of file from the read callback function instead of zero.  It will
soon enforce its use by asserting on a returned length of zero.

It is undocumented which error code should be returned by the callback
if it is passed a NULL buffer pointer or zero length, so MOC will pass
back AVERROR(EINVAL) for now and revise that if the need arises.