Maint: Migrate audio decoding to FFmpeg/LibAV's send/receive API.
While the new API is a better approach, it does not integrate well
with MOC's existing decoder structure.  Emulating the existing (and
now deprecated) avcodec_decode_audio4() using the new send/receive
functions offers the least disruptive approach, but is not the ideal
one.

The significance of the new function's return codes is unclear and
the handling of them may have to be revisited on a case-by-case basis
as usage identifies any remaining issues.